+++ build_test/lib/src/test_builder.dart
@@ -140,6 +140,7 @@ Future<TestBuilderResult> testBuilder(
   Resolvers? resolvers,
   TestReaderWriter? readerWriter,
   AssetGraph? assetGraph,
+  bool verbose = false,
   bool enableLowResourceMode = false,
   bool performCleanup = true,
 }) async {
@@ -156,6 +157,7 @@ Future<TestBuilderResult> testBuilder(
     resolvers: resolvers,
     readerWriter: readerWriter,
     assetGraph: assetGraph,
+    verbose: verbose,
     enableLowResourceMode: enableLowResourceMode,
     performCleanup: performCleanup,
   );
@@ -189,6 +191,7 @@ Future<TestBuilderResult> testBuilders(
   bool testingBuilderConfig = true,
   TestReaderWriter? readerWriter,
   AssetGraph? assetGraph,
+  bool verbose = false,
   bool enableLowResourceMode = false,
   bool performCleanup = true,
 }) {
@@ -230,6 +233,7 @@ Future<TestBuilderResult> testBuilders(
     appliesBuilders: appliesBuildersToFactories,
     testingBuilderConfig: testingBuilderConfig,
     readerWriter: readerWriter,
+    verbose: verbose,
     enableLowResourceMode: enableLowResourceMode,
     performCleanup: performCleanup,
   );
@@ -292,6 +296,9 @@ Future<TestBuilderResult> testBuilders(
 /// Optionally pass [readerWriter] to set the filesystem that will be used
 /// during the build. Before the build, [sourceAssets] will be written to it.
 ///
+/// Optionally pass [verbose], which acts like the command line flag: it enables
+/// info logging from builders.
+///
 /// Optionally pass [enableLowResourceMode], which acts like the command
 /// line flag; in particular it disables file caching.
 ///
@@ -316,6 +323,7 @@ Future<TestBuilderResult> testBuilderFactories(
   bool testingBuilderConfig = true,
   TestReaderWriter? readerWriter,
   AssetGraph? assetGraph,
+  bool verbose = false,
   bool enableLowResourceMode = false,
   bool performCleanup = true,
 }) async {
@@ -365,6 +373,7 @@ Future<TestBuilderResult> testBuilderFactories(
 
   buildLog.configuration = buildLog.configuration.rebuild((b) {
     b.onLog = onLog;
+    b.verbose = verbose;
   });
   resolvers ??=
       packageConfig == null && enabledExperiments.isEmpty
@@ -472,6 +481,7 @@ Future<TestBuilderResult> testBuilderFactories(
     // ignore: invalid_use_of_visible_for_testing_member
     buildOptions: BuildOptions.forTests(
       enableLowResourcesMode: enableLowResourceMode,
+      verbose: verbose,
     ),
     testingOverrides: testingOverrides,
   );
@@ -505,9 +515,23 @@ Future<TestBuilderResult> testBuilderFactories(
 }
 
 class TestBuilderResult {
+  @Deprecated('Use `succeeded`, `outputs` and `errors` instead.')
   final BuildResult buildResult;
   final TestReaderWriter readerWriter;
   final AssetGraph? assetGraph;
+
+  bool get succeeded => buildResult.status == BuildStatus.success;
+
+  /// The files that were written.
+  ///
+  /// If the build was an incremental build, outputs of the previous build that
+  /// are reused are _not_ included here.
+  BuiltList<AssetId> get outputs => buildResult.outputs;
+
+  /// The errors that were shown.
+  ///
+  /// If the build failed there must be at least one error.
+  BuiltList<String> get errors => buildResult.errors;
 
   TestBuilderResult({
     required this.buildResult,