+++ build_test/CHANGELOG.md
@@ -1,7 +1,14 @@
-## 3.4.1-wip
+## 3.5.0-wip
 
-- Use `build_runner` 2.8.1-wip.
-- Use `build` 4.0.1-wip.
+- Improve `TestBuilderResult`: add `succeeded`, `outputs` and `errors`.
+  Deprecate `buildResult` in favor of these new members.
+- Add `verbose` to `testBuilders` and related methods. Like the command line
+  flag it enables info logging from builders.
+
+## 3.4.1
+
+- Use `build_runner` 2.9.0.
+- Use `build` 4.0.1.
 
 ## 3.4.0